Try to remove yet more machine dependencies

+++ rubicon/util/checkstat.c
@@ -27,8 +27,9 @@ int main(int argc, char **argv)
     printf("%d %d %d %d %d %d %d %d %d ",
            s.st_dev, s.st_ino, s.st_mode, s.st_nlink,
            s.st_uid,  s.st_gid, s.st_rdev, s.st_size, s.st_blksize);
-    printf("%d %d %d %d\n",
-           s.st_blocks, s.st_atime, s.st_mtime, s.st_ctime);
+    printf("%lld %ld %ld %ld\n",
+           (long long)s.st_blocks, (long)s.st_atime, (long)s.st_mtime,
+           (long)s.st_ctime);
   }
   else
     return -1;
